+++ activerecord/lib/active_record/alert.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module ActiveRecord
+  module Alert
+    def alert
+      errors.full_messages.to_sentence
+    end
+  end
+end